+++ tests/test_daemon.py
@@ -3,10 +3,7 @@ import os
 import pytest
 import vaping
 import vaping.daemon
-
-
-def test_init():
-    pass
+import vaping.config
 
 
 def test_plugin_context():
@@ -21,8 +18,22 @@ def test_plugin_context():
     ctx.config.data['1'] = 'three'
     assert data != ctx.config.data
 
-def test_empty_config(data_dir):
+
+def test_empty_config_dir(data_dir):
     config_dir = os.path.join(data_dir, 'config', 'empty')
+
     with pytest.raises(ValueError) as excinfo:
         daemon = vaping.daemon.Vaping(config_dir=config_dir)
     assert 'no plugins specified' in str(excinfo)
+
+
+def test_empty_config_dict():
+    with pytest.raises(ValueError) as excinfo:
+        daemon = vaping.daemon.Vaping(config={})
+    assert 'config was not specified' in str(excinfo)
+
+
+def test_empty_config_object():
+    with pytest.raises(ValueError) as excinfo:
+        daemon = vaping.daemon.Vaping(config=vaping.Config())
+    assert 'config was not specified' in str(excinfo)